docs : 청첩장 컴포넌트 폴더링
