+++ Wedit-Client/src/components/WeddingInvitationPage/gallery.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import theme from '../styles/theme';
+import theme from '../../styles/theme';
 
 const Gallery = () => {
 	return (
