+++ Wedit-Client/src/components/WeddingInvitationPage/guest-book-modal.jsx
@@ -1,11 +1,10 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import theme from '../styles/theme';
+import theme from '../../styles/theme';
 
-const RSVPModal = ({ isVisible, onClose }) => {
+const GuestBookModal = ({ isVisible, onClose }) => {
 	const [name, setName] = useState('');
-	const [phone, setPhone] = useState('');
-	const [guestNum, setguestNum] = useState(0);
+	const [content, setContent] = useState('');
 
 	// ModalContainer 밖인 ModalWrapper 클릭 시 모달창 닫힘
 	const handleWrapperClick = (e) => {
@@ -15,14 +14,14 @@ const RSVPModal = ({ isVisible, onClose }) => {
 		}
 	};
 
-	// 이름, 연락처, 추가 인원 선택 여부 확인
-	const isButtonActive = name.trim() !== '' && phone.trim() !== '' && guestNum !== 0;
+	// 이름과 내용 입력 여부 확인
+	const isButtonActive = name.trim() !== '' && content.trim() !== '';
 
 	return (
 		isVisible && (
 			<ModalWrapper onClick={handleWrapperClick}>
 				<ModalContainer>
-					<TitleSpan>참석의사 전달하기</TitleSpan>
+					<TitleSpan>방명록</TitleSpan>
 					<ModalBox>
 						<NameInput
 							type="text"
@@ -30,26 +29,12 @@ const RSVPModal = ({ isVisible, onClose }) => {
 							value={name}
 							onChange={(e) => setName(e.target.value)}
 						/>
-						<PhoneInput
-							placeholder="연락처를 입력해 주세요"
-							value={phone}
-							onChange={(e) => setPhone(e.target.value)}
+						<ContentInput
+							placeholder="내용을 입력해 주세요"
+							value={content}
+							onChange={(e) => setContent(e.target.value)}
 						/>
-						<GuestNumDropdown
-							value={guestNum === 0 ? '' : guestNum}
-							onChange={(e) => setguestNum(Number(e.target.value))}>
-							<option value="" disabled hidden>
-								추가 인원
-							</option>
-							<option>외 0명</option>
-							<option>외 1명</option>
-							<option>외 2명</option>
-							<option>외 3명</option>
-							<option>외 4명</option>
-							<option>외 5명 이상</option>
-						</GuestNumDropdown>
-
-						<SubmitButton isActive={isButtonActive}>전달 하기</SubmitButton>
+						<SubmitButton isActive={isButtonActive}>작성 완료</SubmitButton>
 					</ModalBox>
 				</ModalContainer>
 			</ModalWrapper>
@@ -57,7 +42,7 @@ const RSVPModal = ({ isVisible, onClose }) => {
 	);
 };
 
-export default RSVPModal;
+export default GuestBookModal;
 
 const ModalWrapper = styled.div`
 	position: fixed;
@@ -74,7 +59,7 @@ const ModalWrapper = styled.div`
 
 const ModalContainer = styled.div`
 	width: 80.8rem;
-	height: 70.1rem;
+	height: 78.7rem;
 	border-radius: 4rem;
 	background: #fff;
 	display: flex;
@@ -84,7 +69,7 @@ const ModalBox = styled.div`
 	display: flex;
 	flex-direction: column;
 	align-items: center;
-	gap: 4.5rem;
+	gap: 3rem;
 `;
 
 const TitleSpan = styled.p`
@@ -93,7 +78,7 @@ const TitleSpan = styled.p`
 	font-style: normal;
 	line-height: 106.667%;
 	letter-spacing: -0.057rem;
-	margin: 8.3rem 49.1rem 6.5rem 10.7rem;
+	margin: 8.3rem 62.4rem 6.5rem 10.7rem;
 `;
 
 const NameInput = styled.input`
@@ -101,7 +86,7 @@ const NameInput = styled.input`
 	border: 0.1rem solid #a8a8a8;
 	background: #fff;
 	display: flex;
-	width: 59.3rem;
+	width: 60.1rem;
 	height: 3.2rem;
 	align-items: flex-start;
 	gap: 1rem;
@@ -109,50 +94,34 @@ const NameInput = styled.input`
 	color: #808080;
 	font-weight: ${theme.font.semibold.fontWeight};
 	font-size: ${theme.fontSize.xlarge};
+	font-family: Pretendard;
 	font-style: normal;
 	line-height: 133.333%;
 	letter-spacing: -0.0456rem;
-	padding-left: 3.1rem;
+	padding-left: 2.5rem;
 	padding-top: 2.8rem;
 	padding-bottom: 2.8rem;
 `;
 
-const PhoneInput = styled.input`
+const ContentInput = styled.textarea`
 	border-radius: 1.6rem;
 	border: 0.1rem solid #a8a8a8;
 	background: #fff;
 	display: flex;
-	width: 59.3rem;
-	height: 3.2rem;
+	width: 60.1rem;
+	height: 28.2rem;
 	align-items: flex-start;
 	gap: 1rem;
 	flex-shrink: 0;
 	color: #808080;
 	font-weight: ${theme.font.semibold.fontWeight};
 	font-size: ${theme.fontSize.xlarge};
+	font-family: Pretendard;
 	font-style: normal;
 	line-height: 133.333%;
 	letter-spacing: -0.0456rem;
-	padding-left: 3.1rem;
-	padding-top: 2.8rem;
-	padding-bottom: 2.8rem;
-`;
-
-const GuestNumDropdown = styled.select`
-	border-radius: 1.6rem;
-	border: 0.1rem solid #a8a8a8;
-	background: #fff;
-	display: flex;
-	width: 63.3rem;
-	height: 8.8rem;
-	color: #808080;
-	font-weight: ${theme.font.semibold.fontWeight};
-	font-size: ${theme.fontSize.xlarge};
-	font-style: normal;
-	line-height: 133.333%;
-	letter-spacing: -0.0456rem;
-	padding-left: 2.8rem;
-	padding-right: 2.8rem;
+	padding-left: 2.5rem;
+	padding-top: 3rem;
 `;
 
 const SubmitButton = styled.button`
